requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==42.0.5
cssselect2==0.7.0
dj-database-url==2.1.0
Django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-extensions==3.2.1
django-filter==23.2
django-model-utils==4.3.1
django-multiselectfield==0.1.12
django-otp==1.2.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-timestampedmodel==0.1.0
django-timezone-field==6.1.0
django-user-activity-log2==0.0.25
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-yasg==1.21.5
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==20.1.0
html5lib==1.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.7
lxml==5.2.2
mailchimp-marketing==3.0.80
MarkupSafe==2.1.5
multidict==6.0.5
num2words==0.5.13
numpy==1.26.3
oscrypto==1.3.0
packaging==23.2
pandas==2.2.2
pdfkit==1.0.0
pillow==10.3.0
postgres==4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycparser==2.22
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT
PyMySQL==1.1.0
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
redis==4.5.5
reportlab==4.2.0
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
setuptools==70.0.0
six==1.16.0
SQLAlchemy==1.3.13
sqlparse==0.4.4
stripe==8.2.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.16
yarl==1.9.4
zope.event==5.0
zope.interface==6.3
